Bank Marketing
This project is an exploratory data analysis (EDA) of the Bank Marketing dataset from Kaggle. The purpose of the project is to gain insights and understand the relationships between various factors that influence the success of a marketing campaign for a bank. The analysis is done using Python and several libraries, including pandas and matplotlib, to perform data manipulation and visualization. The dataset contains information about clients of a Portuguese bank who were contacted for a marketing campaign. The goal is to determine which factors are most important in predicting whether the client will subscribe to the bank's product. The end result of the project is a report that summarizes the findings and provides insights that can be used to guide future marketing efforts.